changed precision of center function

+++ Computer_Vision/Demo1/Demo1.py
@@ -119,8 +119,8 @@ def find_center(corners):
         marker_corners = outline.reshape((4,2))
         
         # Compute the center of the marker
-        center_x = int(np.mean(marker_corners[:, 0]))
-        center_y = int(np.mean(marker_corners[:, 1]))
+        center_x = np.mean(marker_corners[:, 0])
+        center_y = np.mean(marker_corners[:, 1])
 
     return (center_x, center_y)
 
@@ -179,7 +179,7 @@ def detect_marker_and_angle():
             center = find_center(corners)
 
             # Mark the center of the marker on the frame (mostly for debuging purposes)
-            cv2.circle(frame_undistorted, center, 3, (0, 255, 0), -1)
+            cv2.circle(frame_undistorted, int(center), 3, (0, 255, 0), -1)
 
             # Calculate the angle of the marker relative to the camera's center
             newAngle = findPhi(center[0], cameraMatrix[0,2], cameraMatrix[0,0])
